# Conflicts:
#	src/modules/charts/components/ResourceDataVizList.vue

+++ src/modules/user/components/DashboardPage.vue
@@ -14,17 +14,19 @@
         :actionButtonCallback="catchSigninClick"
       />
     </div>
-
-    <div v-if="user" class="dashboardGrid">
+    <div v-if="user"
+         class="dashboardGrid">
       <div class="topBoard mt-6 mt-md-0">
         <IntroductionCard
           :userName="user.fullName"
           :introText="userDashboardConfig.introText"
           :feedbackText="userDashboardConfig.feedbackText"
           :oldDashboardUrl="oldDashboardUrl"
-          :createClickCallback="canCreateDatasets ? createClickCallback : null"
+          :createClickCallback="
+            organizationsStore.canCreateDatasets ? createClickCallback : null
+          "
           :editingClickCallback="editingClickCallback"
-          :editingDatasetName="lastEditedDataset"
+          :editingDatasetName="lastDataset"
           :currentLocalDataset="
             hasLocalDataset ? currentLocalDataset : undefined
           "
@@ -41,6 +43,7 @@
           :nameInitials="nameInitials"
           :organizationRoles="organizationRoles"
           :isCollaborator="isCollaborator"
+          @organizationClick="catchOrganizationClick"
         />
 
         <FlipLayout
@@ -249,7 +252,7 @@
  * @author Dominik Haas-Artho
  *
  * Created at     : 2020-07-14 14:18:32
- * Last modified  : 2020-10-13 12:49:34
+ * Last modified  : 2025-10-02 13:16:20
  */
 
 import { mapState, mapGetters } from 'vuex';
@@ -279,8 +282,9 @@ import {
   USER_SIGNIN_PATH,
   METADATADETAIL_PAGENAME,
   METADATAEDIT_PAGENAME,
-  METADATA_CREATION_PATH,
   METADATA_CREATION_PAGENAME,
+  ORGANIZATIONS_PAGENAME,
+  WORKFLOW_PAGENAME,
 } from '@/router/routeConsts';
 
 import { useOrganizationsStore } from '@/modules/organizations/store/organizationsStorePinia';
@@ -300,9 +304,6 @@ import {
   getUserOrganizationRoleMap,
   hasOrganizationRoles,
   isMember,
-  USER_ROLE_ADMIN,
-  USER_ROLE_EDITOR,
-  USER_ROLE_SYSTEM_ADMIN,
 } from '@/factories/userEditingValidations';
 
 import UserNotFound1 from '@/modules/user/assets/UserNotFound1.jpg';
@@ -319,8 +320,6 @@ import {
 
 import { getPreviewDatasetFromLocalStorage } from '@/factories/userCreationFactory';
 
-import { METADATA_TITLE_PROPERTY } from '@/factories/metadataConsts';
-
 import { loadRouteTags } from '@/factories/stringFactory';
 import categoryCards from '@/store/categoryCards';
 
@@ -333,6 +332,7 @@ import EditUserProfile from '@/modules/user/components/edit/EditUserProfile.vue'
 import FlipLayout from '@/components/Layouts/FlipLayout.vue';
 
 import { getMetadataVisibilityState } from '@/factories/publicationFactory';
+import { useDatasetWorkflowStore } from '@/modules/workflow/datasetWorkflow';
 
 const IntroductionCard = defineAsyncComponent(
   () => import('@/components/Cards/IntroductionCard.vue'),
@@ -346,6 +346,8 @@ const NotificationCard = defineAsyncComponent(
 const UserOrganizationInfo = defineAsyncComponent(
   () => import('@/components/Cards/UserOrganizationInfo.vue'),
 );
+
+const workflowStore = useDatasetWorkflowStore();
 
 export default {
   name: 'DashboardPage',
@@ -396,6 +398,9 @@ export default {
       'updatingTags',
       'metadatasContent',
     ]),
+    lastDataset() {
+      return this.lastEditedDataset;
+    },
     userDashboardConfig() {
       return this.config?.userDashboardConfig || {};
     },
@@ -407,6 +412,9 @@ export default {
     },
     datasetCreationActive() {
       return this.userEditMetadataConfig?.datasetCreationActive || false;
+    },
+    newWorkflowActive() {
+      return this.userEditMetadataConfig?.newWorkflowActive || false;
     },
     loading() {
       return (
@@ -535,21 +543,21 @@ export default {
 
       return roles;
     },
-    canCreateDatasets() {
-      const roles = this.organizationRoles;
+    // canCreateDatasets() {
+    //   const roles = this.organizationRoles;
 
-      if (roles) {
-        const matchedRole = roles.filter(
-          (r) =>
-            r.role === USER_ROLE_EDITOR ||
-            r.role === USER_ROLE_ADMIN ||
-            r.role === USER_ROLE_SYSTEM_ADMIN,
-        );
-        return matchedRole.length > 0;
-      }
+    //   if (roles) {
+    //     const matchedRole = roles.filter(
+    //       (r) =>
+    //         r.role === USER_ROLE_EDITOR ||
+    //         r.role === USER_ROLE_ADMIN ||
+    //         r.role === USER_ROLE_SYSTEM_ADMIN,
+    //     );
+    //     return matchedRole.length > 0;
+    //   }
 
-      return false;
-    },
+    //   return false;
+    // },
     isCollaborator() {
       return this.collaboratorDatasets?.length > 0;
     },
@@ -580,9 +588,14 @@ export default {
       return this.user?.fullName?.split(' ')[1] || '';
     },
     hasLocalDataset() {
+      if (this.newWorkflowActive) {
+        const localDataset = this.workflowStore.localStorageService.dataset;
+        return !!localDataset;
+      }
+
       const localStorageData = getPreviewDatasetFromLocalStorage();
 
-      // eslint-disable-next-line no-unused-expressions
+      // eslint-disable-next-line @typescript-eslint/no-unused-expressions
       this.localDatasetUpdateCount;
 
       const properties = Object.keys(localStorageData);
@@ -590,14 +603,15 @@ export default {
       return properties.length > 0;
     },
     currentLocalDataset() {
-      const localStorageData = getPreviewDatasetFromLocalStorage();
+      let localDataset = this.newWorkflowActive ? this.workflowStore.localStorageService.dataset : getPreviewDatasetFromLocalStorage();
 
-      const localDataset = {
-        title: localStorageData[METADATA_TITLE_PROPERTY]
-          ? localStorageData[METADATA_TITLE_PROPERTY]
+      localDataset = {
+        ...localDataset,
+        title: localDataset.title
+          ? localDataset.title
           : '[Untitled Dataset]',
-        subTitle: localStorageData.description,
-        tags: localStorageData.keywords,
+        subTitle: localDataset.description,
+        tags: localDataset.keywords,
         flatLayout: true,
       };
 
@@ -707,7 +721,13 @@ export default {
     },
     createClickCallback() {
       if (this.datasetCreationActive) {
-        this.$router.push({ path: METADATA_CREATION_PATH, query: '' });
+
+        const name = this.newWorkflowActive ? WORKFLOW_PAGENAME : METADATA_CREATION_PAGENAME;
+
+        this.$router.push({
+          name,
+          query: '',
+        });
       } else {
         window.open(`${this.ckanDomain}${this.createCKANUrl}`, '_blank');
       }
@@ -726,12 +746,22 @@ export default {
         // this.catchEditingClick(this.lastEditedDataset);
       }
     },
-    catchEditingClick(selectedDataset) {
+    catchEditingClick(selectedDatasetId) {
+
+      let name = METADATAEDIT_PAGENAME;
+      const params = {
+        metadataid: selectedDatasetId,
+      }
+
+      if (this.newWorkflowActive) {
+        name = WORKFLOW_PAGENAME;
+        params.id = selectedDatasetId;
+        delete params.metadataid;
+      }
+
       this.$router.push({
-        name: METADATAEDIT_PAGENAME,
-        params: {
-          metadataid: selectedDataset,
-        },
+        name,
+        params,
         query: {
           backPath: this.$route.fullPath,
         },
@@ -777,8 +807,19 @@ export default {
       }
     },
     catchLocalCardClick() {
+      const name = this.newWorkflowActive ? WORKFLOW_PAGENAME : METADATA_CREATION_PAGENAME;
+      const params = {};
+
+      if (this.newWorkflowActive) {
+        params.id = this.currentLocalDataset.id;
+      }
+
       this.$router.push({
-        name: METADATA_CREATION_PAGENAME,
+        name,
+        params,
+        query: {
+          backPath: this.$route.fullPath,
+        },
       });
     },
     catchClearLocalStorage() {
@@ -788,7 +829,11 @@ export default {
           'This dataset is not stored on the server yet! Are you sure you want to delete it?',
         callback: () => {},
         cancelCallback: () => {
-          localStorage.clear();
+          if (this.newWorkflowActive) {
+            this.workflowStore.clearLocalStorage();
+          } else {
+            localStorage.clear();
+          }
 
           // increase this number to force the computed property to recalulate because the
           // localstorage is cleared. Localstorage isn't reactive so the computed prop
@@ -797,6 +842,14 @@ export default {
         },
         confirmText: 'Keep Dataset',
         cancelText: 'Delete Dataset',
+      });
+    },
+    catchOrganizationClick(organization) {
+      this.$router.push({
+        name: ORGANIZATIONS_PAGENAME,
+        params: {
+          organization,
+        },
       });
     },
   },
@@ -857,6 +910,7 @@ export default {
     ],
     localDatasetUpdateCount: 0,
     mdiRefresh,
+    workflowStore,
   }),
   components: {
     MetadataList,
